Converted public/js/main.js to an .mjs file that imports the new spookyMeter module as an ES6 module.
