+++ views/layouts/index.pug
@@ -9,4 +9,4 @@ head
 a#fork-me(href='https://github.com/goibon/spooky-meter')
   img(style='position: absolute; top: 0; left: 0; border: 0;' src='https://camo.githubusercontent.com/8b6b8ccc6da3aa5722903da7b58eb5ab1081adee/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f6c6566745f6f72616e67655f6666373630302e706e67' alt='Fork me on GitHub' data-canonical-src='https://s3.amazonaws.com/github/ribbons/forkme_left_orange_ff7600.png')
 script(type='text/javascript' src='node_modules/pulltorefreshjs/dist/pulltorefresh.min.js')
-script(type='text/javascript' src='js/main.js')
+script(type='text/javascript' src='js/main.mjs')
